requirements.txt
altair
pandas
streamlit
google-generativeai
numpy
scikit-learn
python-dotenv